requirements.txt
django==3.1
psycopg2==2.8.5
djangorestframework==3.12.1
beautifulsoup4==4.9.3